delete member table to transform to user
