--- src/Controller/MemberController.php
@@ -1,81 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use App\Entity\Member;
-use App\Form\MemberType;
-use App\Repository\MemberRepository;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Attribute\Route;
-
-#[Route('/member')]
-final class MemberController extends AbstractController
-{
-    #[Route(name: 'app_member_index', methods: ['GET'])]
-    public function index(MemberRepository $memberRepository): Response
-    {
-        return $this->render('member/index.html.twig', [
-            'members' => $memberRepository->findAll(),
-        ]);
-    }
-
-    #[Route('/new', name: 'app_member_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
-    {
-        $member = new Member();
-        $form = $this->createForm(MemberType::class, $member);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->persist($member);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('app_member_index', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->render('member/new.html.twig', [
-            'member' => $member,
-            'form' => $form,
-        ]);
-    }
-
-    #[Route('/{id}', name: 'app_member_show', methods: ['GET'])]
-    public function show(Member $member): Response
-    {
-        return $this->render('member/show.html.twig', [
-            'member' => $member,
-        ]);
-    }
-
-    #[Route('/{id}/edit', name: 'app_member_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Member $member, EntityManagerInterface $entityManager): Response
-    {
-        $form = $this->createForm(MemberType::class, $member);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-
-            return $this->redirectToRoute('app_member_index', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->render('member/edit.html.twig', [
-            'member' => $member,
-            'form' => $form,
-        ]);
-    }
-
-    #[Route('/{id}', name: 'app_member_delete', methods: ['POST'])]
-    public function delete(Request $request, Member $member, EntityManagerInterface $entityManager): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$member->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($member);
-            $entityManager->flush();
-        }
-
-        return $this->redirectToRoute('app_member_index', [], Response::HTTP_SEE_OTHER);
-    }
-}
