--- src/Entity/Member.php
@@ -1,170 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use App\Repository\MemberRepository;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-
-#[ORM\Entity(repositoryClass: MemberRepository::class)]
-#[ORM\Table(name: '`member`')]
-class Member
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
-
-    #[ORM\Column(length: 255)]
-    private ?string $firstName = null;
-
-    #[ORM\Column(length: 255)]
-    private ?string $lastName = null;
-
-    #[ORM\Column(length: 255)]
-    private ?string $phone = null;
-
-    #[ORM\Column(length: 255)]
-    private ?string $mail = null;
-
-    #[ORM\Column]
-    private ?bool $administrator = null;
-
-    #[ORM\Column]
-    private ?bool $active = null;
-
-    /**
-     * @var Collection<int, Output>
-     */
-    #[ORM\OneToMany(targetEntity: Output::class, mappedBy: 'organisator', orphanRemoval: true)]
-    private Collection $outputs;
-
-    #[ORM\ManyToOne(inversedBy: 'members')]
-    #[ORM\JoinColumn(nullable: false)]
-    private ?site $site = null;
-
-    public function __construct()
-    {
-        $this->outputs = new ArrayCollection();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getFirstName(): ?string
-    {
-        return $this->firstName;
-    }
-
-    public function setFirstName(string $firstName): static
-    {
-        $this->firstName = $firstName;
-
-        return $this;
-    }
-
-    public function getLastName(): ?string
-    {
-        return $this->lastName;
-    }
-
-    public function setLastName(string $lastName): static
-    {
-        $this->lastName = $lastName;
-
-        return $this;
-    }
-
-    public function getPhone(): ?string
-    {
-        return $this->phone;
-    }
-
-    public function setPhone(string $phone): static
-    {
-        $this->phone = $phone;
-
-        return $this;
-    }
-
-    public function getMail(): ?string
-    {
-        return $this->mail;
-    }
-
-    public function setMail(string $mail): static
-    {
-        $this->mail = $mail;
-
-        return $this;
-    }
-
-    public function isAdministrator(): ?bool
-    {
-        return $this->administrator;
-    }
-
-    public function setAdministrator(bool $administrator): static
-    {
-        $this->administrator = $administrator;
-
-        return $this;
-    }
-
-    public function isActive(): ?bool
-    {
-        return $this->active;
-    }
-
-    public function setActive(bool $active): static
-    {
-        $this->active = $active;
-
-        return $this;
-    }
-
-    /**
-     * @return Collection<int, Output>
-     */
-    public function getOutputs(): Collection
-    {
-        return $this->outputs;
-    }
-
-    public function addOutput(Output $output): static
-    {
-        if (!$this->outputs->contains($output)) {
-            $this->outputs->add($output);
-            $output->setOrganisator($this);
-        }
-
-        return $this;
-    }
-
-    public function removeOutput(Output $output): static
-    {
-        if ($this->outputs->removeElement($output)) {
-            // set the owning side to null (unless already changed)
-            if ($output->getOrganisator() === $this) {
-                $output->setOrganisator(null);
-            }
-        }
-
-        return $this;
-    }
-
-    public function getSite(): ?site
-    {
-        return $this->site;
-    }
-
-    public function setSite(?site $site): static
-    {
-        $this->site = $site;
-
-        return $this;
-    }
-}
