--- src/Form/MemberType.php
@@ -1,36 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Member;
-use App\Entity\site;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class MemberType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('firstName')
-            ->add('lastName')
-            ->add('phone')
-            ->add('mail')
-            ->add('administrator')
-            ->add('active')
-            ->add('site', EntityType::class, [
-                'class' => site::class,
-                'choice_label' => 'id',
-            ])
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => Member::class,
-        ]);
-    }
-}
