--- templates/member/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit Member{% endblock %}
-
-{% block body %}
-    <h1>Edit Member</h1>
-
-    {{ include('member/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('app_member_index') }}">back to list</a>
-
-    {{ include('member/_delete_form.html.twig') }}
-{% endblock %}
