--- templates/member/index.html.twig
@@ -1,45 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Member index{% endblock %}
-
-{% block body %}
-    <h1>Member index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>FirstName</th>
-                <th>LastName</th>
-                <th>Phone</th>
-                <th>Mail</th>
-                <th>Administrator</th>
-                <th>Active</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for member in members %}
-            <tr>
-                <td>{{ member.id }}</td>
-                <td>{{ member.firstName }}</td>
-                <td>{{ member.lastName }}</td>
-                <td>{{ member.phone }}</td>
-                <td>{{ member.mail }}</td>
-                <td>{{ member.administrator ? 'Yes' : 'No' }}</td>
-                <td>{{ member.active ? 'Yes' : 'No' }}</td>
-                <td>
-                    <a href="{{ path('app_member_show', {'id': member.id}) }}">show</a>
-                    <a href="{{ path('app_member_edit', {'id': member.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="8">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('app_member_new') }}">Create new</a>
-{% endblock %}
