--- templates/member/new.html.twig
@@ -1,11 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}New Member{% endblock %}
-
-{% block body %}
-    <h1>Create new Member</h1>
-
-    {{ include('member/_form.html.twig') }}
-
-    <a href="{{ path('app_member_index') }}">back to list</a>
-{% endblock %}
