--- templates/member/show.html.twig
@@ -1,46 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Member{% endblock %}
-
-{% block body %}
-    <h1>Member</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>Id</th>
-                <td>{{ member.id }}</td>
-            </tr>
-            <tr>
-                <th>FirstName</th>
-                <td>{{ member.firstName }}</td>
-            </tr>
-            <tr>
-                <th>LastName</th>
-                <td>{{ member.lastName }}</td>
-            </tr>
-            <tr>
-                <th>Phone</th>
-                <td>{{ member.phone }}</td>
-            </tr>
-            <tr>
-                <th>Mail</th>
-                <td>{{ member.mail }}</td>
-            </tr>
-            <tr>
-                <th>Administrator</th>
-                <td>{{ member.administrator ? 'Yes' : 'No' }}</td>
-            </tr>
-            <tr>
-                <th>Active</th>
-                <td>{{ member.active ? 'Yes' : 'No' }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('app_member_index') }}">back to list</a>
-
-    <a href="{{ path('app_member_edit', {'id': member.id}) }}">edit</a>
-
-    {{ include('member/_delete_form.html.twig') }}
-{% endblock %}
